% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_file_chunked.R
\name{append_beta_se_to_eqtlgen_cis_eqtl}
\alias{append_beta_se_to_eqtlgen_cis_eqtl}
\title{Append columns for beta and standard error values to the eQTLGen cis-eQTL
dataset.}
\usage{
append_beta_se_to_eqtlgen_cis_eqtl(
  eqtlgen_cis_eqtl_filepath,
  eqtlgen_cis_eqtl_maf_filepath,
  outfile_path,
  chunk_size = 10000
)
}
\arguments{
\item{eqtlgen_cis_eqtl_filepath}{Character. Path to}

\item{eqtlgen_cis_eqtl_maf_filepath}{Character. Path
to\href{https://molgenis26.gcc.rug.nl/downloads/eqtlgen/cis-eqtl/README_cis}{cis-eQTL
file}
\href{https://molgenis26.gcc.rug.nl/downloads/eqtlgen/cis-eqtl/README_allele_frequency}{cis_eQTL
mean allele frequency file}}

\item{outfile_path}{Character. Path to outfile. An error is raised if a file
already exists at this location.}

\item{chunk_size}{The number of rows to include in each chunk}
}
\description{
Calculates beta and standard error values for the eQTLGen eQTLGen
\href{https://molgenis26.gcc.rug.nl/downloads/eqtlgen/cis-eqtl/README_cis}{cis-eQTL}
dataset from Z-score, sample size and allele frequency (these values are in a
\href{https://molgenis26.gcc.rug.nl/downloads/eqtlgen/cis-eqtl/README_allele_frequency}{
separate file}) values.
}
\seealso{
\code{\link{process_file_chunked}},
\code{\link{mutate_beta_se_from_zscore_eqtlgen_cis_eqtl}}
}
