% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_se_from_zscore_maf.R
\name{beta_se_from_zscore}
\alias{beta_se_from_zscore}
\title{Get beta or standard error from z-score and allele frequency}
\usage{
beta_se_from_zscore(df, zscore, maf, sample_n, beta_se)
}
\arguments{
\item{df}{A dataframe containing the required input values}

\item{zscore}{Colname from \code{df} containing zscores}

\item{maf}{Colname from \code{df} containing allele frequencies}

\item{sample_n}{Colname from \code{df} containing sample size}

\item{beta_se}{Character. Must be either \code{"beta"} or \code{"se"}}
}
\value{
A numerical vector of either beta or standard error values
}
\description{
Take a dataframe containing Z-scores, allele frequencies and sample sizes and
returns a vector of either beta or standard error values (specified by the
\code{beta_se} argument)
}
