% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtlgentools.R
\docType{package}
\name{eqtlgentools}
\alias{eqtlgentools}
\title{eqtlgentools: Tools for working with eQTLGen data}
\description{
The main purpose of this package is to create beta and standard error values
in the cis-eQTL data from Z-scores, allele frequencies and sample sizes.
}
