% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_se_from_zscore_maf.R
\name{mutate_beta_se_from_zscore}
\alias{mutate_beta_se_from_zscore}
\title{Append columns for beta and standard error values to a dataframe}
\usage{
mutate_beta_se_from_zscore(df, zscore, maf, sample_n, beta_se)
}
\arguments{
\item{df}{A dataframe containing the required input values}

\item{zscore}{Colname from \code{df} containing zscores}

\item{maf}{Colname from \code{df} containing allele frequencies}

\item{sample_n}{Colname from \code{df} containing sample size}

\item{beta_se}{Character. Must be either \code{"beta"} or \code{"se"}}
}
\value{
A dataframe with appended \code{beta} and \code{se} columns
}
\description{
Appends columns with beta and standard error values to a dataframe containing
columns for Z-scores, allele frequencies and sample sizes.
}
\seealso{
\code{\link{beta_se_from_zscore}}
}
