% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_se_from_zscore_maf.R
\name{mutate_beta_se_from_zscore_eqtlgen_cis_eqtl}
\alias{mutate_beta_se_from_zscore_eqtlgen_cis_eqtl}
\title{Append columns for beta and standard error values to a dataframe of eQTLGen
cis-eQTL data}
\usage{
mutate_beta_se_from_zscore_eqtlgen_cis_eqtl(
  eqtlgen_cis_eqtl_df,
  eqtlgen_cis_eqtl_maf_df
)
}
\arguments{
\item{eqtlgen_cis_eqtl_df}{Dataframe of eQTLGen cis-eQTL data}

\item{eqtlgen_cis_eqtl_maf_df}{eQTLGen cis-eQTL mean allele frequency data}
}
\value{
A dataframe with appended \code{beta} and \code{se} columns
}
\description{
A wrapper around \code{\link{mutate_beta_se_from_zscore}}. Works specifically
for the eQTLGen
\href{https://molgenis26.gcc.rug.nl/downloads/eqtlgen/cis-eqtl/README_cis}{cis-eQTL}
and
\href{https://molgenis26.gcc.rug.nl/downloads/eqtlgen/cis-eqtl/README_allele_frequency}{cis_eQTL
mean allele frequency} datasets.
}
